Reformat the README to use more lists
After looking through some of the READMEs for projects from "A Bestiary
of Single-File Implementations of Programming Languages][0]," I decided
I wanted to try reformatting this README to use more lists, include code
samples, and have shorter instructions for how to run the compiler and
interpreter.

Other changes:

Add a link to "A Bestiary of Single-File Implementations of Programming
Languages."

Switch from absolute to relative links for in-repo items.

[0]: https://github.com/marcpaq/b1fipl