Adjust project title to be less ambiguous
The word "exercise" in the original title was somewhat misleading
because it made it sound like the project might be an exercise to be
completed instead of an already completed toy implementation.

Other changes:

[Rename the default Git branch to "main."][0]

Add a subtitle to mention which file contains the compiler and which
contains the interpreter.

Change "examples" to the more descriptive "language examples."

Use a full sentence for the note about the starting memory address in
the first code example.  I think it's easier to read as a full sentence.

[0]: https://github.com/github/renaming